export core generators as an NPM package
Renamed index.js -> server.js for starting the server
Renamed maker.js -> index.js for importing the generator

Added a compile step so package can be used in pre-ES6 environments
